refactoring into multiple submodules for easier testing and development; initial class-ification of buffer_streams, but not testing new functionality yet.
