+++ pescador/__init__.py
@@ -3,6 +3,8 @@
 
 from .util import *
 from .zmq_stream import *
-from .pescador import *
+from .mux import *
+from .buffered import *
+from .core import *
 
 from .version import version as __version__
